+++ Application/GroupChats/Commands/ChangeGroupChatImage/ChangeGroupChatImageCommandHandler.cs
